enhancement: delete event in event page
